Fix issue when user try to run `no mained` code
User can write code outside the main function. The execution of that
code is undefined, and we don't want in massa to allow that code.

Add a runtime error in the case that we want to run the code before the
initialisation by wasmer of the environment. Send the error `Try to
execute code outside the main function` to the client.